added speechrecognition to predict tricks in gameactivity and added some tests to playertest class

+++ app/src/test/java/com/example/mi_b_wizard/GameTest.java
@@ -127,12 +127,12 @@ public class GameTest {
         Player p3 = new Player("Abel");
 
         p.setMadeTrick(3);
-        p.setPredictedTrick(3);
+        p.setPredictedTrick((byte)3);
         p1.setMadeTrick(2);
-        p1.setPredictedTrick(0);
+        p1.setPredictedTrick((byte)0);
         p2.setMadeTrick(1);
-        p2.setPredictedTrick(1);
-        p3.setPredictedTrick(0);
+        p2.setPredictedTrick((byte)1);
+        p3.setPredictedTrick((byte)0);
         p3.setMadeTrick(0);
 
         testGame.addPlayerToPlayers(p);
